changed the jquery scrolling
I changed the scrolling. I created a single function that handles the
scrolling between scenes. Each click (except the tower) now just makes
a call to a single scroll function. This way we do not have to have
repetative code of the scrolling on each click - they can all just use
the same function. The tower still has it's own scrolling code because
it scrolls at a slower pace than the others.